Refactor tests suite, add tests for share consents
