+++ source/Assets/Plugins/Didomi/Tests/PurposeAndVendorListsTestsSuite.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b86d23502ed544d079979c72b1007e66
+guid: f3006ce4c7f1d4538aa47a690f8beacd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
