+++ source/Assets/Plugins/Didomi/Tests/ShareConsentsTestsSuite.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 894359db9e57d4742a948e1aa685e76c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
